+++ packages/web3-core/test/unit/web3_config.test.ts
@@ -33,7 +33,7 @@ const defaultConfig = {
 		useRpcCallSpecification: false,
 	},
 	handleRevert: false,
-	contractDataInputFill: 'input',
+	contractDataInputFill: 'data',
 	maxListenersWarningThreshold: 100,
 	transactionBlockTimeout: 50,
 	transactionConfirmationBlocks: 24,
@@ -104,6 +104,22 @@ describe('Web3Config', () => {
 			});
 		},
 	);
+
+	it('set default chain error', () => {
+		const obj = new MyConfigObject();
+
+		obj.setConfig({
+			// @ts-expect-error incorrect object
+			defaultCommon: {
+				baseChain: 'mainnet',
+			},
+		});
+		expect(() => {
+			obj.defaultChain = 'test';
+		}).toThrow(
+			'Web3Config chain doesnt match in defaultHardfork mainnet and common.hardfork test',
+		);
+	});
 
 	it('Updating transactionPollingInterval should update transactionReceiptPollingInterval and transactionConfirmationPollingInterval', () => {
 		const obj = new MyConfigObject();