+++ CHANGELOG.md
@@ -2226,6 +2226,7 @@ If there are any bugs, improvements, optimizations or any new feature proposal f
 -   Dependencies updated
 
 ## [4.2.2]
+
 ### Added
 
 #### web3-core
@@ -2253,4 +2254,72 @@ If there are any bugs, improvements, optimizations or any new feature proposal f
 
 -	Will populate `data` for transactions in contract for metamask provider instead of `input` (#6534)
 
+## [4.3.0]
+
+### Changed
+
+#### web3-core
+
+-	Web3config `contractDataInputFill` has been defaulted to `data`, istead of `input`. (#6622)
+
+#### web3-eth-contracts
+
+-	By default, contracts will fill `data` instead of `input` within method calls (#6622)
+
+### Added
+
+
+#### web3-utils
+
+-   `SocketProvider` now contains public function `getPendingRequestQueueSize`, `getSentRequestsQueueSize` and `clearQueues` (#6479)
+-   Added `safeDisconnect` as a `SocketProvider` method to disconnect only when request queue size and send request queue size is 0 (#6479) 
+-   Add `isContractInitOptions` method (#6555)
+
+#### web3
+
+-   Added methods (privateKeyToAddress, parseAndValidatePrivateKey, and privateKeyToPublicKey) to web3.eth.accounts (#6620)
+
+### Fixed
+
+#### web3-rpc-methods
+
+-   Fix web3-types import #6590 (#6589)
+
+#### web3-utils
+
+-   Fix unecessary array copy when pack encoding (#6553)
+
 ## [Unreleased]
+
+### Added
+
+#### web3
+
+
+#### web3-eth
+
+-   Catch `TransactionPollingTimeoutError` was added to send transaction events (#6623)
+
+#### web3-utils
+
+-   `SocketProvider` now contains public function `getPendingRequestQueueSize`, `getSentRequestsQueueSize` and `clearQueues` (#6479)
+-   Added `safeDisconnect` as a `SocketProvider` method to disconnect only when request queue size and send request queue size is 0 (#6479) 
+-   Add `isContractInitOptions` method (#6555)
+
+### Changed
+
+#### web3-core
+
+
+#### web3-eth-contract
+
+
+### Fixed
+
+#### web3-rpc-methods
+
+-   Fix web3-types import #6590 (#6589)
+
+#### web3-utils
+
+-   Fix unecessary array copy when pack encoding (#6553)